http_server: nest panes + activities under windows and split handlers per file
