+++ daemon/http_server/src/handlers.rs
@@ -1,7 +1,5 @@
-pub mod activities;
 pub mod health;
 pub mod index;
-pub mod panes;
 pub mod sessions;
 pub mod windows;
